Switch to zipped portable git
7z could not be extracted by the portable 7z.exe (neither r/a)
portable git as from the official git site is an PE archive
which is not supported by the 7zr.exe we ship.
Use mingit which is zipped, which we can extract using tar
or powershell. Tar is supported on windows10 version from jan2019
and later and powershell is supported on any windows10 (if not
disabled by the user). This should be sufficiently strong
assumptions.

Signed-off-by: Florian Esser <florian.esser@tngtech.com>